+++ src/neuroconv/converters/__init__.py
@@ -19,6 +19,9 @@ from ..datainterfaces.ophys.brukertiff.brukertiffconverter import (
     BrukerTiffMultiPlaneConverter,
     BrukerTiffSinglePlaneConverter,
 )
+from ..datainterfaces.fiber_photometry.guppy.tdtfiberphotometryguppyconverter import (
+    TDTFiberPhotometryGuppyConverter,
+)
 from ..datainterfaces.ophys.miniscope.miniscopeconverter import MiniscopeConverter
 
 converter_list = [
@@ -32,4 +35,5 @@ converter_list = [
     MiniscopeConverter,
     SortedRecordingConverter,
     SortedSpikeGLXConverter,
+    TDTFiberPhotometryGuppyConverter,
 ]